redshift/dataenv/requirement.txt
beautifulsoup4==4.11.1
boto3==1.24.82
botocore==1.27.82
cdata==0.1.9
certifi==2022.9.24
charset-normalizer==2.1.1
greenlet==1.1.3
idna==3.4
jieba==0.42.1
jmespath==1.0.1
numpy==1.23.3
pandas==1.5.0
petl==1.7.11
psycopg2-binary==2.9.3
py4j==0.10.9.5
pypi==2.1
pyspark==3.3.0
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
urllib3==1.26.12
wikipedia==1.4.0
xlrd==2.0.1
xlwt==1.3.0
